+++ packages/react/src/components/Button/styles.tsx
@@ -1,8 +1,7 @@
 import { tv } from 'tailwind-variants'
-import { cn } from '~/utils/cn'
 
 export const button = tv({
-  base: cn([
+  base: [
     'flex items-center justify-center',
     'font-sans font-medium',
     'rounded-full',
@@ -12,7 +11,7 @@ export const button = tv({
     'focus-visible:outline-none',
     'disabled:opacity-32',
     'enabled:active:ring-1',
-  ]),
+  ],
   variants: {
     variant: {
       primary: [
@@ -22,38 +21,38 @@ export const button = tv({
         'enabled:active:border-neutral-900 enabled:active:ring-neutral-900',
       ],
       'primary-inverse': [
-        'border-white bg-white text-primary-500 ',
+        'border-white bg-white text-primary-500',
         'enabled:hover:border-neutral-500 enabled:hover:bg-neutral-500',
         'focus-visible:!border-primary-500',
         'enabled:active:border-primary-500 enabled:active:ring-primary-500',
       ],
       secondary: [
-        'border-primary-500 bg-transparent text-primary-500 ',
+        'border-primary-500 bg-transparent text-primary-500',
         'enabled:hover:border-primary-600 enabled:hover:text-primary-600',
         'focus-visible:!border-neutral-900',
         'enabled:active:border-neutral-900 enabled:active:ring-neutral-900',
       ],
       'secondary-inverse': [
-        'border-white bg-transparent text-white ',
+        'border-white bg-transparent text-white',
         'enabled:hover:border-neutral-500 enabled:hover:text-neutral-500',
         'focus-visible:!border-primary-500',
         'enabled:active:border-primary-500 enabled:active:ring-primary-500',
       ],
-
       cta: [
-        'border-alert-500 bg-alert-500 text-white ',
+        'border-alert-500 bg-alert-500 text-white',
         'enabled:active:hover:bg-alert-500/90',
         'focus-visible:border-neutral-900',
         'enabled:active:border-neutral-900 enabled:active:ring-neutral-900',
       ],
       destructive: [
-        'border-error-500 bg-error-500 text-white ',
+        'border-error-500 bg-error-500 text-white',
         'enabled:active:hover:bg-error-500/90',
         'focus-visible:border-neutral-900',
         'enabled:active:border-neutral-900 enabled:active:ring-neutral-900',
       ],
     },
     size: {
+      '32': ['min-h-[2rem] w-fit min-w-[2rem] p-0.5', 'text-base'],
       sm: ['gap-2 px-4 py-2', 'text-xs leading-none'],
       md: ['gap-2 px-6 py-3', 'text-base leading-none'],
       lg: ['gap-4 px-8 py-4', 'text-base leading-none'],